third_party/blink/web_tests/external/wpt/trusted-types/trusted-types-svg-script-set-href-expected.txt

This is a testharness.js-based test.
[FAIL] Assign string to SVGScriptElement.href.baseVal.
  promise_test: Unhandled rejection with value: object "TypeError: Failed to set the 'data' property on 'HTMLObjectElement': This document requires 'TrustedScriptURL' assignment."
[FAIL] Assign string to non-attached SVGScriptElement.href via setAttribute.
  promise_test: Unhandled rejection with value: object "TypeError: Failed to set the 'data' property on 'HTMLObjectElement': This document requires 'TrustedScriptURL' assignment."
[FAIL] Assign string to attached SVGScriptElement.href via setAttribute.
  promise_test: Unhandled rejection with value: object "TypeError: Failed to set the 'data' property on 'HTMLObjectElement': This document requires 'TrustedScriptURL' assignment."
[FAIL] Assign string to SVGScriptElement.className.baseVal.
  Failed to set the 'baseVal' property on 'SVGAnimatedString': This document requires 'TrustedScriptURL' assignment.
[FAIL] Assign string to SVGScriptElement.href.baseVal  w/ default policy.
  assert_equals: expected "SVGScriptElement href" but got "SVGAnimatedString baseVal"
[FAIL] Assign string to non-attached SVGScriptElement.href via setAttribute w/ default policy.
  assert_equals: expected "SVGScriptElement href" but got "Element setAttribute"
[FAIL] Assign string to attached SVGScriptElement.href via setAttribute w/ default policy.
  assert_equals: expected "SVGScriptElement href" but got "Element setAttribute"
[FAIL] Assign string to SVGScriptElement.className.baseVal w/ default policy.
  assert_equals: expected "SVGScriptElement href" but got "SVGAnimatedString baseVal"
Harness: the test ran to completion.
